# lib/api-zod/src/generated/types/reading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Reading {
  id: number;
  sensor_id: string;
  timestamp: Date;
  humedad?: number | null;
  temperatura?: number | null;
  ec?: number | null;
  bateria?: number | null;
  senal?: number | null;
}