# lib/api-zod/src/generated/types/sensorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SensorInput {
  id_sensor: string;
  nombre_zona: string;
  umbral_humedad_min: number;
  umbral_humedad_max: number;
  umbral_ec_max: number;
}