# lib/api-zod/src/generated/types/sensorSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Reading } from "./reading";
import type { Sensor } from "./sensor";
import type { SensorSummaryStatus } from "./sensorSummaryStatus";

export interface SensorSummary {
  sensor: Sensor;
  latestReading?: Reading | null;
  status: SensorSummaryStatus;
  alerts: string[];
}